Fixes trac:4140: Move r30333 to v1.7 branch (Cleanup warnings for use of)
---svn-pre-commit-ignore-below---

r30333 [[BR]]
Cleanup warnings for use of void* - requires intermediate cast to uintptr_t. Thanks to Paul Hargrove for reporting it

cmr=v1.7.4:reviewer=jsquyres

This commit was SVN r30355.

The following SVN revision numbers were found above:
  r30333 --> open-mpi/ompi@2cf4862b49b2c422bef7edde43e493de9b0c245b

The following Trac tickets were found above:
  Ticket 4140 --> https://svn.open-mpi.org/trac/ompi/ticket/4140

+++ ompi/mca/btl/openib/btl_openib.c
@@ -269,7 +269,7 @@ static int check_if_device_support_modify_srq(mca_btl_openib_module_t *openib_bt
         return rc;
     }
 
-    sge_elem.addr = (uint64_t) &buff;
+    sge_elem.addr = (uint64_t)((uintptr_t) &buff);
     sge_elem.length = sizeof(buff);
 
     wr1.num_sge = wr2.num_sge = 1;
